+++ lib/filters/nested/filter-group.tsx
@@ -1,5 +1,5 @@
 import { useLensContext } from "@lens2/contexts/lens-context";
-import { ConditionRow } from "@lens2/filters/advanced/condition-row";
+import { ConditionRow } from "@lens2/filters/nested/condition-row";
 import { createEmptyCondition } from "@lens2/filters/logic/initialize-filter";
 import { Button } from "@lens2/shadcn/components/ui/button";
 import {
